Update to FindMarkers
The FindMarkers code has a useful line that allows a user to pass cell names directly, instead of an identity class, to the "ident.1" and "ident.2" parameters. As the code is currently written, if a user passes cell names, an error will be thrown because of "cells.1=WhichCells(object,ident.1)" before the intended line of code is ever reached. I have changed the order of the two lines and added an if/else to fix the bug.