fix for issue #188
Allows AverageExpression() to work even when only one gene is provided. By adding "drop = F" the matrix structure of data.temp is maintained even though it only has one row (default would be for its structure to change from a matrix to a vector), allowing data.temp to be successfully passed as an argument to the apply function.