Remove FLIBS from PKG_LIBS
MacOS X versions of R have compilation issues with fortran code.  In particular, the Makeconf file distributed with R points to a fortran library that either doesn't exist or may be out of date.  The problem is described here:

https://thecoatlessprofessor.com/programming/rcpp-rcpparmadillo-and-os-x-mavericks--lgfortran-and--lquadmath-error/

Installing Seurat from source can therefore lead to a compile error depending on the toolings and versions of R/gfortran installed.

However, Seurat does not have any fortran dependencies and so does not need to link to fortran.  By removing this line, we can avoid compilation errors.