Reuse QR Decomposition in RegressOutResid
The RegressOutResid function repeatedly regresses each gene expression pattern against a fixed set of covariates and calculate the residuals.  When the regression model is the std. linear (lm) model, a separate data.frame is formed and passed to an lm command for each regression.  This lm command in turn repeatedly performs a QR decomposition on the fixed covariates, and then calculates the residuals.

To speed this up, rather than calculate the same QR decomposition for each gene, this change calculates it once and then reuses it to re-calculate the residuals with a fast call to `qr.resid`, without the overhead of a full lm call.