Reduce Memory Usage in RegressOutResid
By default, the code in this function would repeatedly create an empty vector with length equal to the rows of the gene expression matrix.  This led to a large amount of allocations and a very large object created in this function.

This change creates one empty vector that is reused instead of creating multiple unique vectors.  Memory profiling for a use case showed the size of data.resid produced in this loop was dropped from 3.5 GB to 1.2 GB.