Set assay = RNA by default in CreateDimReducObject
No longer require the assay parameter in CreateDimReducObject to be
set, instead have assay default to "RNA" with a warning message.

When running dimensional reduction functions directly on matrices
(e.g. RunUMAP.default), there isn't necessarily a clear tie to an
Assay object and therefore requiring this parameter isn't
necessarily useful. By defaulting to "RNA", users can still proceed
with analyses although in the scenario above, it might be a good
idea to use the assay parameter as a way to log/note what matrix
the dimensional reduction was computed on.