Better error checking for FindNeighbors (#1082)
Add a check in FindNeighbors.default to ensure that the input
matrix has rownames. Also, define as.Graph to ensure any matrix
being converted to a Graph class has row/column names.

It was possible to run FindNeighbors.default on a matrix without
rownames (cell names). This would return a Graph without dimnames
which causes problems in downstream functions like FindClusters.
The as.Graph functions ensure that any new Graph has both row and
column names but I also added a check at the beginning of
FindNeighbors.default to catch this problem before spending the work
to compute the graph.