Improve error checking for DoHeatmap (#1088)
If any of the requested features for the requested assay in the
requested assay are not present, omit them and issue a warning.
If none of the requested features are available, stop with a
message.

Currently, if a user requests features that are not available,
DoHeatmap will fail with a "subscript out of bounds" error.
This can happen for example, if the user is requesting features
that haven't been scaled (more common now due to some default
changes). This will allow the plotting to proceed for those features
that are available.