Fix SCE conversion when no colnames for reducedDim (#1089)
If there are no colnames for the embeddings matrix for a reducedDim,
set the key to the name of the reducedDim.

When converting a SingleCellExperiment object to a Seurat object
if any element in the reducedDim slot has no column names for
the embeddings matrix, the key for the DimReduc will try to be set
to character(0) and error out. Easiest solution is to just set the
key to the name of the reducedDim.