Fix dropping of NA ident cells (#1070)
For the purposes of selecting cells in WhichCells, treat NA as its
own class returned by CellsByIdentities.

If the identity of a cell was NA, WhichCells would never return
that cell name. This was because CellsByIdentities returned a
named list with the elements corresponding to the levels (and
NA is not returned as a level). Therefore, add an element at the
end to capture the NA cells if any exist.