Fix FeaturePlot for features containing (), - (#1135)
If passing a feature containing () or -, SingleDimPlot would try to
evaluate it as an expression, leading to errors. Wrap the feature
name, which gets passed from FeaturePlot to the col.by parameter of
SingleDimPlot, in back ticks to avoid this eval problem. This is
how it was being handled in VlnPlot.