Add option for raster vs tile in heatmaps (#994)
When using geom_raster for heatmaps, some PDF viewers (e.g. Preview)
would cause the plots to come out blurry/smeared. This has been
suggested to be a problem with the interpolation of the raster. To
get around this, geom_tile can be used instead. The downside of
geom_tile is that it can be slower to produce/render and creates
larger files.