Improve error handling for FindAllMarkers (#1123)
Retain messages after tryCatch in FindAllMarkers and display them
at the end of the function.

The tryCatch in FindAllMarkers was just swallowing all error messages
and returning an empty dataframe which made it difficult to see
why the error was occuring in FindMarkers. This was originally put
in to allow for testing of a subset of clusters even if some didn't
meet the minimum criteria for the DE tests. Displaying all errors
at the end is preferrable here because it will still allow a partial
FindAllMarkers to complete but also gives the user info about what
failed.