Allow storing of ordered factors in metadata (#925)
When using the double bracket to assign into metadata, we're using
a switch statement. If the value being assigned is an ordered factor,
the class is returned as a length-two vector c("ordered" "factor"),
which isn't valid for EXPR in switch.

Since metadata is the catch-all option, simply take the first item in
this vector as the value for EXPR.