Rename `percent.mito` to `fraction.mito`
The data range is 0 - 1, not 0 - 100.
Thus, when specifying `percent.mito=1` to ask for 1 %, in fact the
interpretation is in 100 %.

Instead of changing the behaviour, the name is adjusted to break as
little existing code (and in a controlled manner only, resulting in an
error message instead of silently changing the results).

All changes in this commit were introduced by the following command:

```sh
find -type f -not -path ./.git/\* \
  -exec sed -i 's/percent\.mito/fraction.mito/g' "{}" \;
```

This commit fixes #1235.