Fix cut_number issue in AddModuleScore (#1227)
If there were lots of duplicate values (usually arising from genes
with zero average expression), this would cause problems for the
cut_number function leading to the error: "insufficient data values
to produce X bins". To solve this, add a small amount of random noise
to the averages when determining the bins to break ties and allow for
approximately equal sized bins to form.