Fix getEdges() Method of the Network Class
The `getEdges` method was a public method of the `Network` class that allows users to obtain a copy of the Edges of a Network object.  This method/interface was ported over from Java to C++ with a bug in it, though because the `getEdges` method is never called by either the Java or C++ version of the code, this bug had no effect on the current code.  However, to improve the code and maintain it's similarity with the original Java version, it's probably worth fixing this unused method.

The specific bug was that instead of filling an array from `start + value_1` to `start + value_2` with a constant value, it was filled from `start + value_1` to `end + value_2` with a constant value, which leads to unexpected behavior due to an array overflow.