bug fixes for keys with underscores
When a key name with an underscore in the middle is passed to
functions that generate DimReducs (e.g RunUMAP), the key gets
truncated at the first underscore. Additionally, if it ends with
a number (e.g. key.name = "umap_test2"), this causes bugs as the
number at the end can get appended to the dimension number (e.g.
dimension 1 becomes dimension 21).