Move metap to suggests, add install message
The metap package is required for FindConservedMarkers. However,
it depends on the mutoss package which depends on the multtest
package. multtest is now hosted on Bioconductor and is causing
issues for automatic package installation via install.packages.

Therefore, since metap is only used in this one function, move it
to Suggests and prompt the user with install instructions if it
is missing.